requirements.txt
# Core dependencies
doclayout-yolo==0.0.3
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.5.0
Pillow>=9.0.0
pyyaml>=6.0.0
albumentations>=1.3.0
pycocotools>=2.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0
pre-commit>=3.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
